Revert "Sub styled header"
