+++ .gitignore
@@ -71,5 +71,3 @@ yarn-error.log
 
 # Vim Swap Files
 *.swp
-*.swo
-*~
