+++ src/components/header.css
@@ -1,0 +1,127 @@
+/* responsive top bar placing elements at ends */
+.topBar {
+  color: White;
+  display: flex;
+  align-items: center;
+  justify-content: space-around;
+  padding: 1em 2em;
+  border-bottom: 2px solid var(--stormy);
+  margin-bottom: 2rem;
+}
+
+.topBar h1 {
+  color: White;
+  margin: 0;
+  line-height: normal;
+}
+
+.header-small-home-link {
+  color: White;
+  font-weight: 100;
+  text-decoration: none;
+  font-size: 1.25rem;
+}
+.header-small-home-link:hover {
+  color: inherit;
+}
+
+/* CSS for links to GitHub/LinkedIn etc. Not required for mobile mode, but necessary for desktop */
+.header-link-pages {
+  display: none
+}
+
+.header-link-pages:hover {
+  color: inherit;
+}
+
+/* BIG home link */
+.header-big-home-link {
+  display: none
+}
+/* if larger, display big link to home */
+
+
+.header-home-link:hover {
+  color: white
+}
+
+.header-link-list {
+  list-style-type: none;
+  display: flex;
+  flex-direction: row;
+  margin: 0;
+  justify-content: space-between;
+}
+
+.header-link-list li {
+  font-size: 1.25rem;
+  padding: 0 1rem;
+  margin: 0;
+  color: White;
+}
+
+.header-link {
+  color: White;
+  font-weight: 100;
+  text-decoration: none;
+  font-size: 1.25rem;
+}
+
+.header-link:hover {
+  color: inherit;
+}
+
+@media only screen and (min-width: 600px) {
+  .topBar {
+    justify-content: space-between;
+  }
+
+  .header-link-pages {
+  display: block;
+  color: White;
+  font-weight: 100;
+  text-decoration: none;
+  font-size: 1.25rem;
+}
+
+.header-link-list li {
+    padding: 0 0.4rem
+  }
+
+  .header-link-list {
+    width: 100%
+  }
+}
+
+
+@media only screen and (min-width: 768px) {
+  .header-small-home-link {
+    display: none
+  }
+
+
+
+  .header-big-home-link {
+    display: block;
+    color: White;
+    font-weight: 100;
+    letter-spacing: -4px;
+    text-decoration: none;
+  }
+
+  .header-link-pages {
+  display: block;
+  color: White;
+  font-weight: 100;
+  text-decoration: none;
+  font-size: 1.25rem;
+}
+
+
+}
+
+@media only screen and (min-width: 1024px) {
+  .header-link-list li {
+    padding: 0 1rem
+  }
+}
