+++ src/components/layout.js
@@ -7,70 +7,10 @@ The query here gets information about the title of the site, as well as icons to
 import React from 'react'
 import PropTypes from 'prop-types'
 import { StaticQuery, graphql } from 'gatsby'
-import styled, { ThemeProvider, createGlobalStyle } from 'styled-components'
 // import './layout.css'
 
-import Header from './Header/header'
-import Footer from './Footer/footer'
-
-//styled-components theme
-import theme from '../theme'
-
-// CSS Reset
-const GlobalStyle = createGlobalStyle`
-	dy, div, span, applet, object, iframe,
-	h1, h2, h3, h4, h5, h6, p, blockquote, pre,
-	a, abbr, acronym, address, big, cite, code,
-	del, dfn, em, img, ins, kbd, q, s, samp,
-	small, strike, strong, sub, sup, tt, var,
-	b, u, i, center,
-	dl, dt, dd, ol, ul, li,
-	fieldset, form, label, legend,
-	table, caption, tbody, tfoot, thead, tr, th, td,
-	article, aside, canvas, details, embed,
-	figure, figcaption, footer, header, hgroup,
-	main, menu, nav, output, ruby, section, summary,
-	time, mark, audio, video {
-		margin: 0;
-		padding: 0;
-		border: 0;
-		font-size: 100%;
-		font: inherit;
-		vertical-align: baseline;
-	}
-	/* HTML5 display-role reset for older browsers */
-	article, aside, details, figcaption, figure,
-	footer, header, hgroup, main, menu, nav, section {
-		display: block;
-	}
-	/* HTML5 hidden-attribute fix for newer browsers */
-	*[hidden] {
-			display: none;
-	}
-	body {
-		line-height: 1;
-	}
-	ol, ul {
-		list-style: none;
-	}
-	blockquote, q {
-		quotes: none;
-	}
-	blockquote:before, blockquote:after,
-	q:before, q:after {
-		content: '';
-		content: none;
-	}
-	table {
-		border-collapse: collapse;
-		border-spacing: 0;
-	}
-  html {
-    height: 100%;
-    width: 100%;
-    background-color: ${props => props.theme.colors.evening}
-  }
-`
+import Header from './header'
+import Footer from './footer'
 
 const Layout = ({ children }) => (
   <StaticQuery
@@ -93,16 +33,13 @@ const Layout = ({ children }) => (
     `}
     render={data => (
       <>
-        <ThemeProvider theme={theme}>
-						<div className="master-container">
-              <div className="body-container">
-                <GlobalStyle />
-								<Header siteTitle={data.homepageInfo.homepageTitle} />
-								{children}
-								<Footer />
-							</div>
-						</div>
-        </ThemeProvider>
+        <div className="master-container">
+          <div className="body-container">
+            <Header siteTitle={data.homepageInfo.homepageTitle} />
+            {children}
+            <Footer />
+          </div>
+        </div>
       </>
     )}
   />
