+++ src/pages/indexStyling.css
@@ -148,6 +148,64 @@ html {
   justify-content: center;
 }
 
+.footer-icons {
+  list-style-type: none;
+  margin: 0;
+  padding: 0;
+  display: flex;
+  flex-direction: row;
+  flex-wrap: wrap;
+}
+
+.footer-icons li:not(:last-child) {
+  margin-right: 10px;
+}
+
+.footer-icon {
+  height: 15px;
+  width: 15px;
+  margin-bottom: 0px
+}
+
+.page-footer {
+  margin-top: 2rem;
+  border-top: 2px solid var(--stormy);
+  font-weight: 100;
+  font-size: 1em;
+  display: flex;
+  padding: 1em 0;
+  height: 3rem;
+  align-items: center;
+  justify-content: space-around;
+}
+
+@media only screen and (min-width: 768px) {
+
+  .footer-icon {
+  height: 25px;
+  width: 25px;
+  margin-bottom: 0px
+  }
+
+  .page-footer {
+  padding: 1em 2em;
+  height: 3rem;
+  font-size: 1em;
+  align-items: center;
+  justify-content: space-between;
+  }
+}
+
+.footer-info {
+  color: White
+}
+
+@media only screen and (max-width:768px) {
+  .footer-info {
+    display: none
+  }
+}
+
 .index-category-image {
   height: 100%;
   width: 100%;
